Moving to its own package - crypto, and some code cleanup.
