--- src/main/java/gov/cdc/izgateway/security/CryptoConfig.java
@@ -1,16 +1,0 @@
-//package gov.cdc.izgateway.security;
-//
-//import lombok.Getter;
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.stereotype.Component;
-//
-//@Component
-//@Getter
-//public class CryptoConfig {
-//PHIZ_CRYPTO_ENCRYPTION_KEY_SECRET_NAME
-//    @Value("${crypto.encryption.key.aws-secret-name:izgw-dev-password-encryption-key}")
-//    private String encryptionKeySecretName;
-//
-//    @Value("${crypto.encryption.key.region:us-east-1}")
-//    private String awsRegion;
-//}
