--- src/main/java/gov/cdc/izgateway/security/KeyProvider.java
@@ -1,5 +1,0 @@
-package gov.cdc.izgateway.security;
-
-public interface KeyProvider {
-    byte[] loadKey() throws CryptoException;
-}
