+++ src/main/java/gov/cdc/izgateway/security/crypto/KeyProviderBase.java
@@ -1,0 +1,31 @@
+package gov.cdc.izgateway.security.crypto;
+
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+public class KeyProviderBase {
+    private final Set<ByteArrayWrapper> keyHistory = new LinkedHashSet<>();
+
+    public List<byte[]> getAllKeys() {
+        synchronized (TestKeyProvider.class) {
+            return keyHistory.stream()
+                    .map(ByteArrayWrapper::getData)
+                    .collect(Collectors.toList());
+        }
+    }
+
+    public void addKeyToHistory(byte[] keyBytes) {
+        synchronized (TestKeyProvider.class) {
+            keyHistory.add(new ByteArrayWrapper(keyBytes));
+        }
+    }
+
+    // Check if key exists
+    public boolean keyExists(byte[] keyBytes) {
+        synchronized (CryptoSupport.class) {
+            return keyHistory.contains(new ByteArrayWrapper(keyBytes));
+        }
+    }
+}
