+++ src/main/java/gov/cdc/izgateway/security/crypto/TestKeyProvider.java
@@ -1,24 +1,8 @@
-package gov.cdc.izgateway.security;
+package gov.cdc.izgateway.security.crypto;
 
 import java.nio.charset.StandardCharsets;
 
-//public class TestKeyProvider implements KeyProvider {
-//
-//
-//    public TestKeyProvider() {
-//    }
-//
-//    @Override
-//    public byte[] loadKey() throws CryptoException {
-//        // Return "MySecretKeyForTestingPurposes12" + 1, 2, or 3 to make it 32 bytes
-//        // This is a placeholder for testing purposes only.
-//
-//        return "MySecretKeyForTestingPurposes123".getBytes(StandardCharsets.UTF_8);
-//    }
-//}
-
-
-public class TestKeyProvider implements KeyProvider {
+public class TestKeyProvider extends KeyProviderBase implements KeyProvider {
     private static int callCount = 0;
     private static final String BASE_KEY = "MySecretKeyForTestingPurposes12";
