+++ src/main/java/gov/cdc/izgateway/soap/net/SoapMessageConverter.java
@@ -8,7 +8,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
 
-import gov.cdc.izgateway.security.CryptoException;
+import gov.cdc.izgateway.security.crypto.CryptoException;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpHeaders;
